pyproject.toml
[project]
name = "Rag_system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "fitz>=0.0.1.dev2",
    "frontend>=0.0.3",
    "google-genai>=1.38.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai",
    "llama-index>=0.14.2",
    "llama-index-embeddings-google-genai>=0.3.1",
    "llama-index-llms-gemini>=0.6.1",
    "llama-index-vector-stores-qdrant>=0.8.5",
    "numpy>=2.3.3",
    "pydantic>=2.11.9",
    "pymongo>=4.15.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.36.0",
]
